// apps/autocad/code-app/src/generated/models/Enmax_autocadbroadcastsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface Enmax_autocadbroadcastsBase {
  enmax_autocadbroadcastid: string;
  enmax_acdntitle?: string;
  enmax_acdnbody?: string;
  enmax_acdnseverity?: number;
  enmax_acdnstatus?: number;
  enmax_acdnaudience?: string;
  enmax_acdnstartsat?: string;
  enmax_acdnexpiresat?: string;
  enmax_acdnpinned?: boolean;
  enmax_acdnrequiresack?: boolean;
  ownerid: string;
  owneridtype: string;
  statecode: number;
}

export interface Enmax_autocadbroadcasts extends Enmax_autocadbroadcastsBase {
  createdon?: string;
  modifiedon?: string;
  _enmax_acdnauthor_value?: string;
  _createdby_value?: string;
  _modifiedby_value?: string;
}